feat: upgrade router to next pages router, fix many issues according to router
