+++ src/components/detailed-card/styles.module.scss
@@ -1,0 +1,77 @@
+.detailed_Card {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  align-items: center;
+}
+
+.card {
+  position: relative;
+  width: 265px;
+  color: var(--text-color);
+  background:
+    linear-gradient(var(--background-color) 0 0) padding-box,
+    linear-gradient(to top, var(--primary-color), var(--secondary-color)) border-box;
+  border: 5px solid transparent;
+  border-radius: 10px;
+}
+
+.text_container {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  padding: 20px 30px;
+  box-shadow: inset 0 5px 5px 0 var(--gray-lighter);
+}
+
+.card_title {
+  margin: 0;
+  font-size: 18px;
+  font-weight: 800;
+  text-align: center;
+}
+
+.desc {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  margin: 0;
+  padding: 0;
+  list-style-type: none;
+}
+
+.desc_item,
+.type {
+  margin: auto;
+}
+
+.episodes {
+  display: -webkit-box;
+  overflow: hidden;
+  margin: 0 auto;
+  text-align: left;
+  text-overflow: ellipsis;
+  -webkit-line-clamp: 5;
+  line-clamp: 5;
+  -webkit-box-orient: vertical;
+}
+
+.close {
+  font-size: 1.3rem;
+}
+
+.loader {
+  width: 265px;
+  height: 265px;
+  background-image: url("../../assets/images/loader.png");
+  background-position: center;
+  background-size: contain;
+  background-repeat: no-repeat;
+  animation: loader 1s infinite linear;
+}
+
+@keyframes loader {
+  to {
+    transform: rotate(1turn);
+  }
+}
