+++ src/components/pagination/types.ts
@@ -3,6 +3,5 @@ import { type MouseEventHandler } from 'react';
 export interface PaginationProps {
   page: number;
   total: number;
-  setPage: (page: number) => void;
   handleClose: MouseEventHandler;
 }
