+++ src/components/results/Results.tsx
@@ -8,7 +8,7 @@ import { Pagination } from '../pagination/Pagination';
 import styles from './styles.module.scss';
 import type { ResultsProps } from './types';
 
-export function Results({ characters, total, page, setPage, characterID, handleClose }: ResultsProps): ReactNode {
+export function Results({ characters, total, page, characterID, handleClose }: ResultsProps): ReactNode {
   return characters.length ? (
     <div className={characterID !== '/' ? styles.results : styles.no_character} onClick={handleClose}>
       <ul className={styles.characters} onClick={handleClose} role="presentation">
@@ -18,7 +18,7 @@ export function Results({ characters, total, page, setPage, characterID, handleC
           </li>
         ))}
       </ul>
-      <Pagination total={total} page={page} setPage={setPage} handleClose={handleClose} />
+      <Pagination total={total} page={page} handleClose={handleClose} />
     </div>
   ) : (
     <div className={styles.not_found_box} onClick={handleClose}>
