+++ src/components/search/Search.tsx
@@ -1,16 +1,13 @@
 import { useRouter } from 'next/router';
 import { type ReactNode } from 'react';
 
-// import { useNavigate, useSearchParams } from 'react-router-dom';
 import { useLocalStorage } from '@/hooks/useLocalStorage';
 
 import styles from './styles.module.scss';
 import type { SearchProps } from './types';
 
-export function Search({ searchField, character, setCharacter, setPage, loader }: SearchProps): ReactNode {
-  // const [, setPageQuery] = useSearchParams();
+export function Search({ searchField, character, setCharacter, loader }: SearchProps): ReactNode {
   const [, setLs] = useLocalStorage('R&M_search');
-  // const navigate = useNavigate();
   const { replace } = useRouter();
 
   return (
@@ -22,9 +19,6 @@ export function Search({ searchField, character, setCharacter, setPage, loader }
           const searchValue = searchField.current?.value.trim();
           setCharacter({ name: searchValue });
           setLs(searchValue);
-          setPage(1);
-          // setPageQuery({ page: '1' });
-          // navigate('/', { replace: true });
           replace({ pathname: '/', query: { page: '1' } }).catch(() => {});
         }
       }}
