+++ src/components/search/types.ts
@@ -2,6 +2,5 @@ export interface SearchProps {
   searchField: React.RefObject<HTMLInputElement>;
   character: { name: string };
   setCharacter: (character: { name: string }) => void;
-  setPage: (page: number) => void;
   loader: boolean;
 }
