+++ src/pages/_error.tsx
@@ -1,28 +1,13 @@
 import Image from 'next/image';
-import { type ReactNode /* , useEffect, useState */ } from 'react';
+import { type ReactNode } from 'react';
 
-// import { isRouteErrorResponse, useRouteError } from 'react-router-dom';
 import error_img from '@/assets/images/loader.png';
 import { useTheme } from '@/hooks/useTheme';
 
 import styles from './error.module.scss';
 
 export default function ErrorPage({ error }: { error: Error & { digest?: string } }): ReactNode {
-  // const error = useRouteError();
-  // const [errorMessage, setErrorMessage] = useState<string>();
   const { theme } = useTheme();
-
-  /*   useEffect(() => {
-    if (error) {
-      if (isRouteErrorResponse(error)) {
-        setErrorMessage(error.statusText);
-      } else if (error instanceof Error) {
-        setErrorMessage(error.message);
-      } else {
-        setErrorMessage(String(error));
-      }
-    }
-  }, [error]); */
 
   return (
     <main className={theme === 'dark' ? 'main' : 'main light'}>
